Write a function that prints all the elements of a list_t list.

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-	int nodes_n = 0;
+	size_t nodes_n = 0;
 
 	while (h)
 	{
